New and improved tests and service ratings
Improved the functionality of many old tests, but they are still a mess
because I am new to automatic unit-testing, not to mention angular, karma
and jasmine. The RatingService now works almost the way it is supposed to
and the directives for displaying content just need to have css applied to
them and have a way to show stars for ratings.