Backend, slight UI improvements
The backend in PHP is included in the backend directory now. The
notification setting was removed in the app, because there are no
notifications anymore. The Terms of Service and Privacy Policy have been
polished. The Contact page had its first version implemented, which still
has a lot of issues.